moved authmodal examples to designmanual
