+++ packages/designmanual/stories/molecules/TagSelectorExample.tsx
@@ -33,7 +33,6 @@ const dummyData = [
 ];
 
 const Container = styled.div`
-  margin: 10px auto;
   max-width: 600px;
 `;
