+++ packages/ndla-ui/src/Footer/index.ts
@@ -9,6 +9,5 @@
 import Footer from './Footer';
 import { FooterText } from './FooterText';
 import { EditorName } from './EditorName';
-import FooterAuth from './FooterAuth';
 
-export { Footer, FooterText, EditorName, FooterAuth };
+export { Footer, FooterText, EditorName };
