+++ packages/ndla-ui/src/User/index.ts
@@ -6,10 +6,8 @@
  *
  */
 
-import AuthModal from './AuthModal';
 import { UserInfo } from './UserInfo';
 import type { FeideGoGroup, FeideGroup, FeideOrg, FeideUserApiType, FeideMembershipType, FeideUser } from './apiTypes';
 
 export { UserInfo };
 export type { FeideGoGroup, FeideGroup, FeideOrg, FeideUserApiType, FeideMembershipType, FeideUser };
-export default AuthModal;
