+++ packages/ndla-ui/src/index.ts
@@ -89,7 +89,7 @@ export {
 
 export { FilterList, FilterListPhone, FilterButtons } from './Filter';
 
-export { Footer, EditorName, FooterText, FooterAuth } from './Footer';
+export { Footer, EditorName, FooterText } from './Footer';
 
 export {
   Figure,
@@ -124,9 +124,7 @@ export { default as ContentTypeResult } from './Search/ContentTypeResult';
 export { SearchFieldForm } from './Search/SearchFieldForm';
 
 export { default as MastheadSearchModal } from './Masthead/MastheadSearchModal';
-export { default as MastheadAuthModal } from './Masthead/MastheadAuthModal';
 export { UserInfo } from './User';
-export { default as AuthModal } from './User';
 export type { FeideGoGroup, FeideGroup, FeideOrg, FeideUserApiType, FeideMembershipType, FeideUser } from './User';
 
 export { default as CreatedBy } from './CreatedBy';
